requirements.txt
annotated-types==0.7.0
asgiref==3.11.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
Django==5.2.9
django-debug-toolbar==6.1.0
django-environ==0.12.0
django-ninja==1.5.1
django-vite==3.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
huey==2.5.5
idna==3.11
inertia-django==1.2.0
pillow==12.0.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
python-json-logger==4.0.0
requests==2.32.5
sqlparse==0.5.4
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
whitenoise==6.11.0